+++ lm-agent/lm_agent/services/reconciliation.py
@@ -2,22 +2,19 @@
 """
 Reconciliation functionality live here.
 """
-import asyncio
-from typing import Dict, List
+from typing import List
 
-from lm_agent.backend_utils.models import BookingSchema
 from lm_agent.backend_utils.utils import (
     get_all_features_bookings_sum,
-    get_bookings_for_all_jobs,
     get_cluster_configs_from_backend,
-    get_cluster_grace_times,
+    get_cluster_jobs_from_backend,
     make_feature_update,
-    remove_job_by_slurm_job_id,
 )
 from lm_agent.exceptions import LicenseManagerEmptyReportError, LicenseManagerReservationFailure
-from lm_agent.license_report import report
 from lm_agent.logs import logger
 from lm_agent.server_interfaces.license_server_interface import LicenseReportItem
+from lm_agent.services.license_report import report
+from lm_agent.services.clean_jobs_and_bookings import clean_jobs_and_bookings
 from lm_agent.workload_managers.slurm.cmd_utils import (
     get_all_features_cluster_values,
     return_formatted_squeue_out,
@@ -31,81 +28,7 @@ from lm_agent.workload_managers.slurm.reservations import (
 )
 
 
-def get_greatest_grace_time_for_job(grace_times: Dict[int, int], job_bookings: List[BookingSchema]) -> int:
-    """
-    Find the greatest grace_time among the features booked by the given job_id.
-    """
-    greatest_booking_feature = max(
-        job_bookings, default=None, key=lambda job_booking: grace_times[job_booking.feature_id]
-    )
-    if not greatest_booking_feature:
-        return -1
-    return grace_times[greatest_booking_feature.feature_id]
-
-
-def get_running_jobs(squeue_result: List) -> List:
-    return [j for j in squeue_result if j["state"] == "RUNNING"]
-
-
-async def clean_jobs_by_grace_time():
-    """
-    Clean the jobs where running time is greater than the grace_time.
-    """
-    logger.debug("GRACE_TIME START")
-
-    cluster_jobs_bookings = await get_bookings_for_all_jobs()
-
-    formatted_squeue_output = return_formatted_squeue_out()
-    if not formatted_squeue_output:
-        logger.debug("GRACE_TIME no squeue")
-        await clean_jobs(None, cluster_jobs_bookings)
-        logger.debug("GRACE_TIME cleaned bookings that are not in the queue")
-        return
-
-    squeue_result = squeue_parser(formatted_squeue_output)
-    squeue_running_jobs = get_running_jobs(squeue_result)
-
-    grace_times = await get_cluster_grace_times()
-
-    # get the grace_time for each job
-    for job in squeue_running_jobs:
-        slurm_job_id = str(job["job_id"])
-        greatest_grace_time = get_greatest_grace_time_for_job(
-            grace_times, cluster_jobs_bookings.get(slurm_job_id, [])
-        )
-        # if the running_time is greater than the greatest grace_time, delete the booking for it
-        if job["run_time_in_seconds"] > greatest_grace_time and greatest_grace_time != -1:
-            logger.debug(f"GRACE_TIME: {greatest_grace_time}, {slurm_job_id}")
-            await remove_job_by_slurm_job_id(slurm_job_id)
-
-    await clean_jobs(squeue_result, cluster_jobs_bookings)
-
-
-async def clean_jobs(squeue_result, cluster_jobs_bookings):
-    """Clean the jobs that aren't running along with its bookings."""
-    logger.debug("CLEAN_JOBS: start")
-
-    if squeue_result is None:
-        squeue_result = []
-
-    jobs_not_running = [str(job["job_id"]) for job in squeue_result if job["state"] != "RUNNING"]
-    all_jobs_squeue = [str(job["job_id"]) for job in squeue_result]
-    logger.debug("CLEAN_JOBS: after building lists")
-
-    delete_job_call = []
-
-    for slurm_job_id in cluster_jobs_bookings.keys():
-        if slurm_job_id in jobs_not_running or slurm_job_id not in all_jobs_squeue:
-            delete_job_call.append(remove_job_by_slurm_job_id(slurm_job_id))
-
-    if not delete_job_call:
-        logger.debug("CLEAN_JOBS: no need to clean")
-        return
-
-    await asyncio.gather(*delete_job_call)
-
-
-async def create_or_update_reservation(reservation_data):
+async def create_or_update_reservation(reservation_data: str):
     """
     Create the reservation if it doesn't exist, otherwise update it.
     If the reservation cannot be updated, delete it and create a new one.
@@ -128,8 +51,16 @@ async def reconcile():
     """Generate the report and reconcile the license feature token usage."""
     logger.debug("Starting reconciliation")
 
+    # Generate report and update the backend
+    logger.debug("Reconciling licenses in the backend")
+    license_usage_info = await update_features()
+    logger.debug("Backend licenses reconciliated")
+
     # Get cluster data
     configurations = await get_cluster_configs_from_backend()
+
+    # Get cluster jobs
+    jobs = await get_cluster_jobs_from_backend()
 
     # Get feature bookings sum
     all_features_bookings_sum = await get_all_features_bookings_sum()
@@ -137,15 +68,11 @@ async def reconcile():
     # Get license usage from the cluster
     all_features_cluster_value = await get_all_features_cluster_values()
 
-    # Delete bookings for jobs that reached the grace time
-    logger.debug("Cleaning jobs by grace time")
-    await clean_jobs_by_grace_time()
-    logger.debug("Jobs cleaned by grace time")
+    # Get squeue result from cluster
+    squeue_output = squeue_parser(return_formatted_squeue_out())
 
-    # Generate report and update the backend
-    logger.debug("Reconciling licenses in the backend")
-    license_usage_info = await update_features()
-    logger.debug("Backend licenses reconciliated")
+    # Clean jobs and bookings
+    await clean_jobs_and_bookings(configurations, jobs, squeue_output, license_usage_info)
 
     reservation_data = []
